# Parallelize with rayon