--- a/Cargo.toml
+++ b/Cargo.toml
@@ -26,6 +26,7 @@ paw = "1.0"
 structopt = { version = "0.3", features = [ "paw" ] }
 anyhow = "1.0.26"
 glob = "0.3.0"
+rayon = "1.3.0"
 
 [dev-dependencies]
 tempfile = "3.1.0"