I've coerced `&Repository` and `&Config` to be `Send + Sync`.
They are just reading states from the git, so I believe they are "semantically" safe to share between threads.
https://github.com/libgit2/libgit2/blob/master/docs/threading.md#sharing-objects

On my machine, it really does reduces the overall time (i7-4702MQ, 4 cores)
With 16 branches that is generated with this script https://github.com/foriequal0/git-trim/issues/22#issuecomment-587084485
before:
```
git trim --dry-run --no-update  18.94s user 0.90s system 99% cpu 19.957 total
```
after:
```
git trim --dry-run --no-update  19.18s user 0.99s system 257% cpu 7.823 total
```

I think this might be premature optimization though. It hurts the readability of logs. But it is fun to do it.